Move spatial grids to deeper package
