+++ gale-server/src/main/java/org/galemc/gale/util/collection/spatialgrid/despawn/point/CylinderSpatialGrid.java
@@ -1,22 +1,23 @@
 // Generated with AI
 
-package org.galemc.gale.util.collection.spatialgrid;
+package org.galemc.gale.util.collection.spatialgrid.despawn.point;
 
-public final class EllipsoidSpatialGrid extends SpatialGrid {
+public final class CylinderSpatialGrid extends SpatialGrid {
 
     // ---------------- constructor ----------------
 
-    public EllipsoidSpatialGrid(int W, int H, int initialCapacity) {
+    public CylinderSpatialGrid(int W, int H, int initialCapacity) {
         super(W, H, initialCapacity);
     }
 
     // ---------------- geometric tests ----------------
 
-    private boolean pointInEllipsoidSlot(int s, double cxD, double cyD, double czD) {
+    private boolean pointInCylinderSlot(int s, double cxD, double cyD, double czD) {
+        double py = ys[s];
+        if (py < cyD - H || py > cyD + H) return false;
         double dx = xs[s] - cxD;
-        double dy = ys[s] - cyD;
         double dz = zs[s] - czD;
-        return dx * dx * invW2 + dy * dy * invH2 + dz * dz * invW2 <= 1.0;
+        return dx * dx + dz * dz <= W2;
     }
 
     // ---------------- queries ----------------
@@ -32,18 +33,18 @@ public final class EllipsoidSpatialGrid extends SpatialGrid {
                     int idx = bucketIndexForCellCoords(ccx + dx, ccy + dy, ccz + dz);
                     if (idx == -1) continue;
                     if (aabbDirty[idx]) recomputeAABB(idx);
-                    if (!aabbIntersectsEllipsoid(idx, cxD, cyD, czD)) continue;
+                    if (!aabbIntersectsCylinder(idx, cxD, cyD, czD)) continue;
                     int cnt = inlineCount[idx];
                     if (cnt > 0) {
-                        if (inline0[idx] != -1 && pointInEllipsoidSlot(inline0[idx], cxD, cyD, czD)) return inline0[idx];
-                        if (cnt >= 2 && inline1[idx] != -1 && pointInEllipsoidSlot(inline1[idx], cxD, cyD, czD)) return inline1[idx];
-                        if (cnt >= 3 && inline2[idx] != -1 && pointInEllipsoidSlot(inline2[idx], cxD, cyD, czD)) return inline2[idx];
-                        if (cnt >= 4 && inline3[idx] != -1 && pointInEllipsoidSlot(inline3[idx], cxD, cyD, czD)) return inline3[idx];
+                        if (inline0[idx] != -1 && pointInCylinderSlot(inline0[idx], cxD, cyD, czD)) return inline0[idx];
+                        if (cnt >= 2 && inline1[idx] != -1 && pointInCylinderSlot(inline1[idx], cxD, cyD, czD)) return inline1[idx];
+                        if (cnt >= 3 && inline2[idx] != -1 && pointInCylinderSlot(inline2[idx], cxD, cyD, czD)) return inline2[idx];
+                        if (cnt >= 4 && inline3[idx] != -1 && pointInCylinderSlot(inline3[idx], cxD, cyD, czD)) return inline3[idx];
                     }
                     int h = overflowHead[idx];
                     while (h != -1) {
                         int s = ovSlot[h];
-                        if (pointInEllipsoidSlot(s, cxD, cyD, czD)) return s;
+                        if (pointInCylinderSlot(s, cxD, cyD, czD)) return s;
                         h = ovNext[h];
                     }
                 }
@@ -65,34 +66,34 @@ public final class EllipsoidSpatialGrid extends SpatialGrid {
                     int idx = bucketIndexForCellCoords(ccx + dx, ccy + dy, ccz + dz);
                     if (idx == -1) continue;
                     if (aabbDirty[idx]) recomputeAABB(idx);
-                    if (!aabbIntersectsEllipsoid(idx, cxD, cyD, czD)) continue;
+                    if (!aabbIntersectsCylinder(idx, cxD, cyD, czD)) continue;
                     int cnt = inlineCount[idx];
                     if (cnt > 0) {
                         if (inline0[idx] != -1) {
                             int s = inline0[idx];
-                            if (pointInEllipsoidSlot(s, cxD, cyD, czD)) {
-                                double dd = dist2SlotToCenter(s, cxD, cyD, czD);
+                            if (pointInCylinderSlot(s, cxD, cyD, czD)) {
+                                double dd = dist2SlotPlanar(s, cxD, czD);
                                 if (dd < best) { best = dd; bestSlot = s; }
                             }
                         }
                         if (cnt >= 2 && inline1[idx] != -1) {
                             int s = inline1[idx];
-                            if (pointInEllipsoidSlot(s, cxD, cyD, czD)) {
-                                double dd = dist2SlotToCenter(s, cxD, cyD, czD);
+                            if (pointInCylinderSlot(s, cxD, cyD, czD)) {
+                                double dd = dist2SlotPlanar(s, cxD, czD);
                                 if (dd < best) { best = dd; bestSlot = s; }
                             }
                         }
                         if (cnt >= 3 && inline2[idx] != -1) {
                             int s = inline2[idx];
-                            if (pointInEllipsoidSlot(s, cxD, cyD, czD)) {
-                                double dd = dist2SlotToCenter(s, cxD, cyD, czD);
+                            if (pointInCylinderSlot(s, cxD, cyD, czD)) {
+                                double dd = dist2SlotPlanar(s, cxD, czD);
                                 if (dd < best) { best = dd; bestSlot = s; }
                             }
                         }
                         if (cnt >= 4 && inline3[idx] != -1) {
                             int s = inline3[idx];
-                            if (pointInEllipsoidSlot(s, cxD, cyD, czD)) {
-                                double dd = dist2SlotToCenter(s, cxD, cyD, czD);
+                            if (pointInCylinderSlot(s, cxD, cyD, czD)) {
+                                double dd = dist2SlotPlanar(s, cxD, czD);
                                 if (dd < best) { best = dd; bestSlot = s; }
                             }
                         }
@@ -100,8 +101,8 @@ public final class EllipsoidSpatialGrid extends SpatialGrid {
                     int h = overflowHead[idx];
                     while (h != -1) {
                         int s = ovSlot[h];
-                        if (pointInEllipsoidSlot(s, cxD, cyD, czD)) {
-                            double dd = dist2SlotToCenter(s, cxD, cyD, czD);
+                        if (pointInCylinderSlot(s, cxD, cyD, czD)) {
+                            double dd = dist2SlotPlanar(s, cxD, czD);
                             if (dd < best) { best = dd; bestSlot = s; }
                         }
                         h = ovNext[h];
@@ -114,17 +115,17 @@ public final class EllipsoidSpatialGrid extends SpatialGrid {
 
     // ---------------- AABB helpers ----------------
 
-    private boolean aabbIntersectsEllipsoid(int idx, double cxD, double cyD, double czD) {
+    private boolean aabbIntersectsCylinder(int idx, double cxD, double cyD, double czD) {
         if (inlineCount[idx] == 0 && overflowHead[idx] == -1) return false;
         double minx = bMinX[idx], maxx = bMaxX[idx];
         double miny = bMinY[idx], maxy = bMaxY[idx];
         double minz = bMinZ[idx], maxz = bMaxZ[idx];
-        double exMin = cxD - W, exMax = cxD + W;
-        double eyMin = cyD - H, eyMax = cyD + H;
-        double ezMin = czD - W, ezMax = czD + W;
-        if (maxx < exMin || minx > exMax) return false;
-        if (maxy < eyMin || miny > eyMax) return false;
-        if (maxz < ezMin || minz > ezMax) return false;
+        double cxMin = cxD - W, cxMax = cxD + W;
+        double cyMin = cyD - H, cyMax = cyD + H;
+        double czMin = czD - W, czMax = czD + W;
+        if (maxx < cxMin || minx > cxMax) return false;
+        if (maxy < cyMin || miny > cyMax) return false;
+        if (maxz < czMin || minz > czMax) return false;
         return true;
     }
