+++ gale-server/src/main/java/org/galemc/gale/util/collection/spatialgrid/despawn/point/SpatialGrid.java
@@ -1,6 +1,6 @@
 // Generated with AI
 
-package org.galemc.gale.util.collection.spatialgrid;
+package org.galemc.gale.util.collection.spatialgrid.despawn.point;
 
 import net.minecraft.core.BlockPos;
 import java.util.Arrays;
